QTIFingerprintHal: Be quiet!
The FP HAL spams like hell when using the fingerprint reader.
Touching the reader with the correct finger causes 272(!) logcat lines. This slows the wake up significantly...
Since there is no switch built into the HAL to disable this output (as far as strings and grep show me), just shim the whole log function to return nothing.

A good indicator of how much the spamming slows down the wake up are the physical buttons. They light up quite fast but the screen is activated only after all the logging is done.
With this shim in place the screen and buttons light up nearly simultaneously.

Another good indicator is only setting up a lockscreen PIN. It will unlock instantly, where-as lockscreen PIN with a fingerprint registered? It takes four seconds.

Change-Id: Ibdeaf5feb1f208b6ef9482b787ead8f5aefb3fef